[Fledge plugin interface] Fix signatures by replacing 'PLUGIN_HANDLE *' with 'PLUGIN_HANDLE'
* Context
---------

The plugin interface is an interface between the Fledge 'core' using
Python language and the plugin written in C++ [1].

The type of PLUGIN_HANDLE is 'void *'.

* Issue
-------

There is an inconsistency between the plugin interface signatures

 - Reading plugin_poll(PLUGIN_HANDLE *handle);
 - void plugin_register_ingest(PLUGIN_HANDLE *handle, INGEST_CB cb, void *data);
 - void plugin_start(PLUGIN_HANDLE *handle);

 and the 'handle' variable handling inside the method bodies.
 Even in the official documentation, about the South plugins only,
 there is this mistake [2].

At the runtime,
the cast operation between 'void **' and 'MyPlugin*' goes smoothly.

With the recommendations of the 'cpplint' tool,
do not use the C cast
but C++ cast like 'reinterpret_cast', or 'static_cast(),
the issue appears:

'reinterpret_cast' is a cast operation executed at the runtime. It is
strongly recommended not to use it.
'static_cast' is a cast operation realized at the build time, and the
compiler checks the compatibility and consistency of the types.

With the use of 'static_cast', the type provided by the signature method
and the 'MyPlugin *' pointer type are checked and matched.
    `IEC61850 *iec61850 = static_cast<IEC61850 *>(*handle);`
But during the runtime, there is a Segmentation fault!

The signature of plugin interfaces contains a mistake.

* Fix
-----

It must be a 'PLUGIN_HANDLE' object instead of 'PLUGIN_HANDLE *'.

And this mistake does not concern the North plugins,
as we can see on the documentation page: [3]

[1] fledge/C/services/south-plugin-interfaces/python/python_plugin_interface.cpp
[2] https://fledge-iot.readthedocs.io/en/develop/plugin_developers_guide/03_south_C_plugins.html
[3] https://fledge-iot.readthedocs.io/en/develop/plugin_developers_guide/04_north_plugins.html#c-c-plugins

Signed-off-by: Mikael Bourhis <mikael.bourhis@smile.fr>